requirements.txt
h5py>=2.10.0
matplotlib>=3.2.2
monai>=0.6.0
nibabel>=3.2.1
numpy>=1.20.3
pydicom>=2.0.0
pytorch-lightning==1.4.5
scikit-image>=0.17.2
scikit-learn>=0.23.2
scipy>=1.5.1
segmentation-models-pytorch>=0.1.2
SimpleITK>=2.0.1
tensorboard>=2.5.0
tqdm>=4.49.0
connected-components-3d>=3.1.2
PyQt5
opencv-python
pandas==1.3.4